requirements.txt
alembic==1.4.3
<<<<<<< HEAD
astroid==2.4.2
blinker==1.4
cffi==1.14.3
=======
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cheroot==8.4.5
>>>>>>> 012ee76fed2f3b9b7d60f85a333209fa90234d27
click==7.1.2
cryptography==2.9.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
flask-talisman==0.7.0
Flask-WTF==0.14.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
<<<<<<< HEAD
lazy-object-proxy==1.4.3
limits==1.5.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
=======
limits==1.5.1
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.5.0
>>>>>>> 012ee76fed2f3b9b7d60f85a333209fa90234d27
pycparser==2.20
pycryptodome==3.9.8
pycryptodomex==3.9.8
pylint==2.6.0
pyOpenSSL==19.1.0
pyotp==2.4.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
<<<<<<< HEAD
toml==0.10.1
=======
urllib3==1.25.10
web.py==0.61
>>>>>>> 012ee76fed2f3b9b7d60f85a333209fa90234d27
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
